refactor: approve with ledger (#194)
